refactor(ta_library): Extract period messages builder, period enum and distance predicate in separated files
