+++ ta_library/src/main/kotlin/com/github/marlonlom/utilities/timeago/TimeAgo.kt
@@ -16,7 +16,6 @@
  */
 package com.github.marlonlom.utilities.timeago
 
-import kotlin.math.abs
 import kotlin.math.roundToLong
 
 /**
@@ -51,241 +50,6 @@ class TimeAgo
  */
 private constructor() {
 
-  /**
-   * The enum Periods.
-   *
-   * @author marlonlom
-   * @version 4.1.0
-   * @since 2.0.0
-   */
-  enum class Periods(
-    /**
-     * The property key.
-     */
-    val propertyKey: String,
-    /**
-     * The predicate.
-     */
-    private val predicate: DistancePredicate,
-  ) {
-
-    NOW(
-      "ml.timeago.now",
-      object : DistancePredicate {
-        override fun validateDistanceMinutes(distance: Long): Boolean = distance in 0L..0.99.toLong()
-      },
-    ),
-    ONE_MINUTE_PAST(
-      "ml.timeago.oneminute.past",
-      object : DistancePredicate {
-        override fun validateDistanceMinutes(distance: Long): Boolean = distance == 1L
-      },
-    ),
-    X_MINUTES_PAST(
-      "ml.timeago.xminutes.past",
-      object : DistancePredicate {
-        override fun validateDistanceMinutes(distance: Long): Boolean = distance in 2..44
-      },
-    ),
-    ABOUT_AN_HOUR_PAST(
-      "ml.timeago.aboutanhour.past",
-      object : DistancePredicate {
-        override fun validateDistanceMinutes(distance: Long): Boolean = distance in 45..89
-      },
-    ),
-    X_HOURS_PAST(
-      "ml.timeago.xhours.past",
-      object : DistancePredicate {
-        override fun validateDistanceMinutes(distance: Long): Boolean = distance in 90..1439
-      },
-    ),
-    ONE_DAY_PAST(
-      "ml.timeago.oneday.past",
-      object : DistancePredicate {
-        override fun validateDistanceMinutes(distance: Long): Boolean = distance in 1440..2519
-      },
-    ),
-    X_DAYS_PAST(
-      "ml.timeago.xdays.past",
-      object : DistancePredicate {
-        override fun validateDistanceMinutes(distance: Long): Boolean = distance in 2520..10079
-      },
-    ),
-    ONE_WEEK_PAST(
-      "ml.timeago.oneweek.past",
-      object : DistancePredicate {
-        override fun validateDistanceMinutes(distance: Long): Boolean = distance in 10080..20159
-      },
-    ),
-    X_WEEKS_PAST(
-      "ml.timeago.xweeks.past",
-      object : DistancePredicate {
-        override fun validateDistanceMinutes(distance: Long): Boolean = distance in 20160..43199
-      },
-    ),
-    ABOUT_A_MONTH_PAST(
-      "ml.timeago.aboutamonth.past",
-      object : DistancePredicate {
-        override fun validateDistanceMinutes(distance: Long): Boolean = distance in 43200..86399
-      },
-    ),
-    X_MONTHS_PAST(
-      "ml.timeago.xmonths.past",
-      object : DistancePredicate {
-        override fun validateDistanceMinutes(distance: Long): Boolean = distance in 86400..525599
-      },
-    ),
-    ABOUT_A_YEAR_PAST(
-      "ml.timeago.aboutayear.past",
-      object : DistancePredicate {
-        override fun validateDistanceMinutes(distance: Long): Boolean = distance in 525600..655199
-      },
-    ),
-    OVER_A_YEAR_PAST(
-      "ml.timeago.overayear.past",
-      object : DistancePredicate {
-        override fun validateDistanceMinutes(distance: Long): Boolean = distance in 655200..914399
-      },
-    ),
-    ALMOST_TWO_YEARS_PAST(
-      "ml.timeago.almosttwoyears.past",
-      object : DistancePredicate {
-        override fun validateDistanceMinutes(distance: Long): Boolean = distance in 914400..1051199
-      },
-    ),
-    X_YEARS_PAST(
-      "ml.timeago.xyears.past",
-      object : DistancePredicate {
-        override fun validateDistanceMinutes(distance: Long): Boolean = (distance / 525600).toFloat().roundToLong() > 1
-      },
-    ),
-    ONE_MINUTE_FUTURE(
-      "ml.timeago.oneminute.future",
-      object : DistancePredicate {
-        override fun validateDistanceMinutes(distance: Long): Boolean = distance.toInt() == -1
-      },
-    ),
-    X_MINUTES_FUTURE(
-      "ml.timeago.xminutes.future",
-      object : DistancePredicate {
-        override fun validateDistanceMinutes(distance: Long): Boolean = distance <= -2 && distance >= -44
-      },
-    ),
-    ABOUT_AN_HOUR_FUTURE(
-      "ml.timeago.aboutanhour.future",
-      object : DistancePredicate {
-        override fun validateDistanceMinutes(distance: Long): Boolean = distance <= -45 && distance >= -89
-      },
-    ),
-    X_HOURS_FUTURE(
-      "ml.timeago.xhours.future",
-      object : DistancePredicate {
-        override fun validateDistanceMinutes(distance: Long): Boolean = distance <= -90 && distance >= -1439
-      },
-    ),
-    ONE_DAY_FUTURE(
-      "ml.timeago.oneday.future",
-      object : DistancePredicate {
-        override fun validateDistanceMinutes(distance: Long): Boolean = distance <= -1440 && distance >= -2519
-      },
-    ),
-    X_DAYS_FUTURE(
-      "ml.timeago.xdays.future",
-      object : DistancePredicate {
-        override fun validateDistanceMinutes(distance: Long): Boolean = distance <= -2520 && distance >= -10079
-      },
-    ),
-    ONE_WEEK_FUTURE(
-      "ml.timeago.oneweek.future",
-      object : DistancePredicate {
-        override fun validateDistanceMinutes(distance: Long): Boolean = distance <= -10080 && distance >= -20159
-      },
-    ),
-    X_WEEKS_FUTURE(
-      "ml.timeago.xweeks.future",
-      object : DistancePredicate {
-        override fun validateDistanceMinutes(distance: Long): Boolean = distance <= -20160 && distance >= -43199
-      },
-    ),
-    ABOUT_A_MONTH_FUTURE(
-      "ml.timeago.aboutamonth.future",
-      object : DistancePredicate {
-        override fun validateDistanceMinutes(distance: Long): Boolean = distance <= -43200 && distance >= -86399
-      },
-    ),
-    X_MONTHS_FUTURE(
-      "ml.timeago.xmonths.future",
-      object : DistancePredicate {
-        override fun validateDistanceMinutes(distance: Long): Boolean = distance <= -86400 && distance >= -525599
-      },
-    ),
-    ABOUT_A_YEAR_FUTURE(
-      "ml.timeago.aboutayear.future",
-      object : DistancePredicate {
-        override fun validateDistanceMinutes(distance: Long): Boolean = distance <= -525600 && distance >= -655199
-      },
-    ),
-    OVER_A_YEAR_FUTURE(
-      "ml.timeago.overayear.future",
-      object : DistancePredicate {
-        override fun validateDistanceMinutes(distance: Long): Boolean = distance <= -655200 && distance >= -914399
-      },
-    ),
-    ALMOST_TWO_YEARS_FUTURE(
-      "ml.timeago.almosttwoyears.future",
-      object : DistancePredicate {
-        override fun validateDistanceMinutes(distance: Long): Boolean = distance <= -914400 && distance >= -1051199
-      },
-    ),
-    X_YEARS_FUTURE(
-      "ml.timeago.xyears.future",
-      object : DistancePredicate {
-        override fun validateDistanceMinutes(distance: Long): Boolean = (distance / 525600).toFloat().roundToLong() < -1
-      },
-    ),
-    ;
-
-    companion object {
-
-      /**
-       * Find by distance minutes periods.
-       *
-       * @param distanceMinutes the distance minutes
-       * @return the periods
-       */
-      fun findByDistanceMinutes(distanceMinutes: Long): Periods? {
-        val values = entries.toTypedArray()
-        for (item in values) {
-          val successful = item.predicate
-            .validateDistanceMinutes(distanceMinutes)
-          if (successful) {
-            return item
-          }
-        }
-        return null
-      }
-    }
-  }
-
-  /**
-   * Interface definition for handling distance validations or periods.
-   *
-   * @author marlonlom
-   * @version 4.1.0
-   * @see Periods
-   *
-   * @since 1.0.0
-   */
-  private interface DistancePredicate {
-    /**
-     * Validate distance minutes boolean.
-     *
-     * @param distance the distance
-     * @return the boolean
-     */
-    fun validateDistanceMinutes(distance: Long): Boolean
-  }
-
   companion object {
 
     /**
@@ -314,138 +78,85 @@ private constructor() {
     private fun buildTimeagoText(resources: TimeAgoMessages, dim: Long): StringBuilder {
       val timeAgo = StringBuilder()
 
-      val foundTimePeriod = Periods.findByDistanceMinutes(dim)
-      if (foundTimePeriod != null) {
-        val periodKey = foundTimePeriod.propertyKey
-        when (foundTimePeriod) {
-          Periods.X_MINUTES_PAST -> timeAgo.append(resources.getPropertyValue(periodKey, dim))
-          Periods.X_HOURS_PAST -> {
-            val hours = (dim / 60f).roundToLong()
-            val xHoursText = handlePeriodKeyAsPlural(
-              resources,
-              "ml.timeago.aboutanhour.past",
-              periodKey,
-              hours.toInt(),
-            )
-            timeAgo.append(xHoursText)
-          }
+      val foundTimePeriod = Periods.findByDistanceMinutes(dim) ?: return timeAgo
 
-          Periods.X_DAYS_PAST -> {
-            val days = (dim / 1440f).roundToLong()
-            val xDaysText = handlePeriodKeyAsPlural(
-              resources,
-              "ml.timeago.oneday.past",
-              periodKey,
-              days.toInt(),
-            )
-            timeAgo.append(xDaysText)
-          }
-
-          Periods.X_WEEKS_PAST -> {
-            val weeks = (dim / 10080f).roundToLong()
-            val xWeeksText = handlePeriodKeyAsPlural(
-              resources,
-              "ml.timeago.oneweek.past",
-              periodKey,
-              weeks.toInt(),
-            )
-            timeAgo.append(xWeeksText)
-          }
-
-          Periods.X_MONTHS_PAST -> {
-            val months = (dim / 43200f).roundToLong()
-            val xMonthsText = handlePeriodKeyAsPlural(
-              resources,
-              "ml.timeago.aboutamonth.past",
-              periodKey,
-              months.toInt(),
-            )
-            timeAgo.append(xMonthsText)
-          }
-
-          Periods.X_YEARS_PAST -> {
-            val years = (dim / 525600f).roundToLong()
-            timeAgo.append(resources.getPropertyValue(periodKey, years))
-          }
-
-          Periods.X_MINUTES_FUTURE -> timeAgo.append(resources.getPropertyValue(periodKey, abs(dim.toFloat())))
-          Periods.X_HOURS_FUTURE -> {
-            val hours1 = abs((dim / 60f).roundToLong())
-            val yHoursText = if (hours1.toInt() == 24) {
-              resources.getPropertyValue("ml.timeago.oneday.future")
-            } else {
-              handlePeriodKeyAsPlural(
-                resources,
-                "ml.timeago.aboutanhour.future",
-                periodKey,
-                hours1.toInt(),
-              )
-            }
-            timeAgo.append(yHoursText)
-          }
-
-          Periods.X_DAYS_FUTURE -> {
-            val days1 = abs((dim / 1440f).roundToLong())
-            val yDaysText = handlePeriodKeyAsPlural(
-              resources,
-              "ml.timeago.oneday.future",
-              periodKey,
-              days1.toInt(),
-            )
-            timeAgo.append(yDaysText)
-          }
-
-          Periods.X_WEEKS_FUTURE -> {
-            val weeks1 = abs((dim / 10080f).roundToLong())
-            val yWeeksText = handlePeriodKeyAsPlural(
-              resources,
-              "ml.timeago.oneweek.future",
-              periodKey,
-              weeks1.toInt(),
-            )
-            timeAgo.append(yWeeksText)
-          }
-
-          Periods.X_MONTHS_FUTURE -> {
-            val months1 = abs((dim / 43200f).roundToLong())
-            val yMonthsText = if (months1.toInt() == 12) {
-              resources.getPropertyValue("ml.timeago.aboutayear.future")
-            } else {
-              handlePeriodKeyAsPlural(
-                resources,
-                "ml.timeago.aboutamonth.future",
-                periodKey,
-                months1.toInt(),
-              )
-            }
-            timeAgo.append(yMonthsText)
-          }
-
-          Periods.X_YEARS_FUTURE -> {
-            val years1 = abs((dim / 525600f).roundToLong())
-            timeAgo.append(resources.getPropertyValue(periodKey, years1))
-          }
-
-          else -> timeAgo.append(resources.getPropertyValue(periodKey))
+      val periodKey = foundTimePeriod.propertyKey
+      when (foundTimePeriod) {
+        Periods.X_MINUTES_PAST -> {
+          timeAgo.append(PeriodMessages.getPastXMinutesText(resources, dim, periodKey))
         }
+
+        Periods.X_HOURS_PAST -> {
+          timeAgo.append(
+            PeriodMessages.getPastXHoursText(resources, dim, periodKey)
+          )
+        }
+
+        Periods.X_DAYS_PAST -> {
+          timeAgo.append(
+            PeriodMessages.getPastXDaysText(resources, dim, periodKey)
+          )
+        }
+
+        Periods.X_WEEKS_PAST -> {
+          timeAgo.append(
+            PeriodMessages.getPastXWeeksText(resources, dim, periodKey)
+          )
+        }
+
+        Periods.X_MONTHS_PAST -> {
+          timeAgo.append(
+            PeriodMessages.getPastXMonthsText(resources, dim, periodKey)
+          )
+        }
+
+        Periods.X_YEARS_PAST -> {
+          timeAgo.append(
+            PeriodMessages.getPastXYearsText(resources, dim, periodKey)
+          )
+        }
+
+        Periods.X_MINUTES_FUTURE -> {
+          timeAgo.append(
+            PeriodMessages.getFutureXMinutesText(resources, dim, periodKey)
+          )
+        }
+
+        Periods.X_HOURS_FUTURE -> {
+          timeAgo.append(
+            PeriodMessages.getFutureXHoursText(resources, dim, periodKey)
+          )
+        }
+
+        Periods.X_DAYS_FUTURE -> {
+          timeAgo.append(
+            PeriodMessages.getFutureXDaysText(resources, dim, periodKey)
+          )
+        }
+
+        Periods.X_WEEKS_FUTURE -> {
+          timeAgo.append(
+            PeriodMessages.getFutureXWeeksText(resources, dim, periodKey)
+          )
+        }
+
+        Periods.X_MONTHS_FUTURE -> {
+          timeAgo.append(
+            PeriodMessages.getFutureXMonthsText(resources, dim, periodKey)
+          )
+        }
+
+        Periods.X_YEARS_FUTURE -> {
+          timeAgo.append(
+            PeriodMessages.getFutureXYearsText(resources, dim, periodKey)
+          )
+        }
+
+        else -> timeAgo.append(resources.getPropertyValue(periodKey))
       }
+
       return timeAgo
     }
-
-    /**
-     * Handle period key as plural string.
-     *
-     * @param resources the resources
-     * @param periodKey the period key
-     * @param value     the value
-     * @return the string
-     */
-    private fun handlePeriodKeyAsPlural(
-      resources: TimeAgoMessages,
-      periodKey: String,
-      pluralKey: String,
-      value: Int,
-    ): String = if (value == 1) resources.getPropertyValue(periodKey) else resources.getPropertyValue(pluralKey, value)
 
     /**
      * Returns the time distance in minutes.
